Reorganize router files
